+++ example/src/App/routes.js
@@ -7,11 +7,12 @@ import SideEffect from '../SideEffect';
 import AsyncSideEffect from '../AsyncSideEffect';
 import LoadingSideEffect from '../LoadingSideEffect';
 import SetNextCall from '../SetNextCall';
+import Selector from '../Selector';
 import Form from '../Form';
 import RegionForm from '../RegionForm';
 import AsyncValidate from '../AsyncValidate';
-import ErrorTodo from '../ErrorTodo';
-import Selector from '../Selector';
+import Error from '../Error';
+import ExtendRegion from '../ExtendRegion';
 
 export default [{
   key: 'home',
@@ -63,10 +64,14 @@ export default [{
   Component: RegionForm,
 }, {
   key: 'AsyncValidate',
-  label: 'Async Validate(TODO)',
+  label: 'Async Validate',
   Component: AsyncValidate,
 }, {
   key: 'Error',
-  label: 'Error(Todo)',
-  Component: ErrorTodo,
+  label: 'Error',
+  Component: Error,
+}, {
+  key: 'ExtendRegion',
+  label: 'Extend Region',
+  Component: ExtendRegion,
 }];