+++ example/src/shared/Card.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Card as RawCard } from 'antd';
+
+const Card = ({ children, ...rest }) => (
+  <RawCard style={{ width: 500, margin: 30 }} {...rest}>{children}</RawCard>
+);
+
+export default Card;