+++ src/region/wrapGet.js
@@ -16,13 +16,13 @@ const formatLoading = (loading, { strictLoading }) => {
 export default (RegionIn) => {
   class Region extends RegionIn {
     getState = () => {
-      const { reducerPath } = this;
+      const { name } = this;
       const { getState } = getStore();
       const state = getState();
-      if (reducerPath === null) {
+      if (name === null) {
         return state || {};
       }
-      return state[reducerPath] || {};
+      return state[name] || {};
     }
 
     getLoading = (path) => {
@@ -42,6 +42,20 @@ export default (RegionIn) => {
       return formatLoading(loadings[path], { strictLoading });
     }
 
+    getFetchTimes = (path) => {
+      const { getState } = this;
+      const { fetchTimes = {} } = getState();
+      if (Array.isArray(path)) {
+        const ans = [];
+        for (let i = 0; i < path.length; i++) {
+          const key = path[i];
+          ans.push(fetchTimes[key]);
+        }
+        return ans;
+      }
+      return fetchTimes[path];
+    }
+
     getResults = (path) => {
       const { getState } = this;
       const { results = {} } = getState();
@@ -56,18 +70,21 @@ export default (RegionIn) => {
       return results[path];
     }
 
-    getFetchTimes = (path) => {
+    getError = (path) => {
       const { getState } = this;
-      const { fetchTimes = {} } = getState();
+      const { errors = {} } = getState();
       if (Array.isArray(path)) {
-        const ans = [];
+        let ans = '';
         for (let i = 0; i < path.length; i++) {
           const key = path[i];
-          ans.push(fetchTimes[key]);
+          const error = errors[key];
+          if (error) {
+            ans += error.message;
+          }
         }
         return ans;
       }
-      return fetchTimes[path];
+      return errors[path];
     }
   }
   return Region;