+++ src/region/wrapSet.js
@@ -7,13 +7,19 @@ export default (RegionIn) => {
      * @param format A function format result to other data structure
      */
     set = (key, result, { format } = {}) => {
-      const { getResults: getSnapshot, SET_RESULT } = this;
+      const { getResults: getSnapshot, private_actionTypes } = this;
+      const { SET } = private_actionTypes;
       const { dispatch } = getStore();
       const snapshot = getSnapshot(key);
 
-      const formattedResult = formatResult({ result, snapshot, key, format });
-      dispatch({ type: SET_RESULT, payload: { key, result: formattedResult } });
-      return formattedResult;
+      try {
+        const formattedResult = formatResult({ result, snapshot, key, format });
+        dispatch({ type: SET, payload: { key, result: formattedResult } });
+        return formattedResult;
+      } catch (error) {
+        dispatch({ type: SET, payload: { key, result: null, error } });
+        return null;
+      }
     }
   }
   return Region;