+++ src/region/wrapSetConfig.js
@@ -1,20 +1,26 @@
 import { setStore } from '../global/store';
+import getActionTypes from '../util/getActionTypes';
 
 export default () => {
   class Region {
     constructor(config) {
       this.private_setConfig({
-        reducerPath: null,
+        name: null,
         expiredTime: 0,
         enableLog: true,
         strictLoading: true,
         silentConnect: false
       });
-      this.private_setConfig(config);
+      if (config !== null && typeof config === 'object') {
+        this.private_setConfig(config);
+      } else {
+        this.private_setConfig({ name: config });
+      }
     }
 
     private_setConfig = (config = {}) => {
       const {
+        name,
         reducerPath,
         expiredTime,
         enableLog,
@@ -22,10 +28,14 @@ export default () => {
         silentConnect
       } = config;
 
+      if (name !== undefined) {
+        this.name = name;
+        this.private_actionTypes = getActionTypes(name);
+      }
       if (reducerPath !== undefined) {
-        this.reducerPath = reducerPath;
-        this.SET_LOADING = reducerPath ? `@${reducerPath}/SET_LOADING` : '@region/SET_LOADING';
-        this.SET_RESULT = reducerPath ? `@${reducerPath}/SET_RESULT` : '@region/SET_RESULT';
+        console.warn('reducerPath is deprecated, use name instead');
+        this.name = reducerPath;
+        this.private_actionTypes = getActionTypes(reducerPath);
       }
       if (expiredTime !== undefined) {
         this.expiredTime = expiredTime;