+++ src/util/__test__/config.test.js
@@ -4,30 +4,34 @@ const { setConfig } = region;
 
 describe('config', () => {
   test('default', () => {
-    const { reducerPath, enableLog, expiredTime, strictLoading, silentConnect, SET_LOADING, SET_RESULT } = region;
-    expect(reducerPath).toBe(null);
+    const { name, enableLog, expiredTime, strictLoading, silentConnect, private_actionTypes } = region;
+    expect(name).toBe(null);
     expect(expiredTime).toBe(0);
     expect(enableLog).toBe(true);
     expect(strictLoading).toBe(true);
     expect(silentConnect).toBe(false);
-    expect(SET_LOADING).toBe('@region/SET_LOADING');
-    expect(SET_RESULT).toBe('@region/SET_RESULT');
+    expect(private_actionTypes).toEqual({
+      LOAD_START: '@region/LOAD',
+      SET: '@region/SET',
+    });
   });
   test('set nothing', () => {
     setConfig();
-    const { reducerPath, enableLog, expiredTime, strictLoading, silentConnect, SET_LOADING, SET_RESULT } = region;
-    expect(reducerPath).toBe(null);
+    const { name, enableLog, expiredTime, strictLoading, silentConnect, private_actionTypes } = region;
+    expect(name).toBe(null);
     expect(expiredTime).toBe(0);
     expect(enableLog).toBe(true);
     expect(strictLoading).toBe(true);
     expect(silentConnect).toBe(false);
-    expect(SET_LOADING).toBe('@region/SET_LOADING');
-    expect(SET_RESULT).toBe('@region/SET_RESULT');
+    expect(private_actionTypes).toEqual({
+      LOAD_START: '@region/LOAD',
+      SET: '@region/SET',
+    });
   });
   test('set enableLog as false', () => {
     setConfig({ enableLog: false });
-    const { reducerPath, enableLog, expiredTime } = region;
-    expect(reducerPath).toBe(null);
+    const { name, enableLog, expiredTime } = region;
+    expect(name).toBe(null);
     expect(expiredTime).toBe(0);
     expect(enableLog).toBe(false);
   });
@@ -35,17 +39,30 @@ describe('config', () => {
     setConfig({
       enableLog: false,
       expiredTime: 30000,
-      reducerPath: 'result',
+      name: 'result',
       strictLoading: false,
       silentConnect: true
     });
-    const { reducerPath, enableLog, expiredTime, strictLoading, silentConnect, SET_LOADING, SET_RESULT } = region;
-    expect(reducerPath).toBe('result');
+    const { name, enableLog, expiredTime, strictLoading, silentConnect, private_actionTypes } = region;
+    expect(name).toBe('result');
     expect(expiredTime).toBe(30000);
     expect(enableLog).toBe(false);
     expect(strictLoading).toBe(false);
     expect(silentConnect).toBe(true);
-    expect(SET_LOADING).toBe('@result/SET_LOADING');
-    expect(SET_RESULT).toBe('@result/SET_RESULT');
+    expect(private_actionTypes).toEqual({
+      LOAD_START: '@result/LOAD',
+      SET: '@result/SET',
+    });
+  });
+  test('deprecated', () => {
+    setConfig({
+      reducerPath: 'deprecated',
+    });
+    const { name, private_actionTypes } = region;
+    expect(name).toBe('deprecated');
+    expect(private_actionTypes).toEqual({
+      LOAD_START: '@deprecated/LOAD',
+      SET: '@deprecated/SET',
+    });
   });
 });