+++ src/util/__test__/logger.test.js
@@ -1,4 +1,4 @@
-import { debug, group, groupError } from '../logger';
+import { debug, group } from '../logger';
 
 const useTrip = () => {
   const trip = [];
@@ -19,19 +19,5 @@ describe('logger', () => {
     const trip = useTrip();
     group('a', 'b', 'c', 'd');
     expect(trip).toEqual(['groupCollapsed', 'debug', 'debug', 'groupEnd']);
-  });
-  test('groupWarn trip', () => {
-    const trip = useTrip();
-    groupError('title', 'e');
-    expect(trip).toEqual(['groupCollapsed', 'error', 'groupEnd']);
-  });
-  test('groupWarn in production', () => {
-    expect(process.env.NODE_ENV).toEqual('test');
-    process.env.NODE_ENV = 'production';
-    const trip = useTrip();
-    groupError('title', 'e');
-    expect(trip).toEqual([]);
-    process.env.NODE_ENV = 'test';
-    expect(process.env.NODE_ENV).toEqual('test');
   });
 });