+++ src/util/__test__/region.js
@@ -3,4 +3,4 @@ import Region from '../../region';
 import { setStore } from '../../global/store';
 
 setStore(createStore(() => {}));
-export const region = new Region({ reducerPath: null });
+export const region = new Region(null);