+++ src/util/formatResult.js
@@ -1,14 +1,7 @@
-import { groupError } from './logger';
-
-export const formatResult = ({ result, snapshot, key, format }) => {
+export const formatResult = ({ result, snapshot, format }) => {
   if (typeof format !== 'function') {
     return result;
   }
-  try {
-    const formattedResult = format(result, snapshot);
-    return formattedResult;
-  } catch (e) {
-    groupError(`Catch an error when format ${key}, return null instead.`, e);
-    return null;
-  }
+  const formattedResult = format(result, snapshot);
+  return formattedResult;
 };