+++ src/util/getActionTypes.js
@@ -1,0 +1,6 @@
+const generator = (name, type) => (name ? `@${name}/${type}` : `@region/${type}`);
+
+export default (name) => ({
+  LOAD_START: generator(name, 'LOAD'),
+  SET: generator(name, 'SET'),
+});