+++ docs/CHANGELOG.md
@@ -1,10 +1,18 @@
 # CHANGELOG
 
+### 0.6.1
+
+- fix loading when several load is called
+
+- new Region('name') & new Region({ name })
+
+- error support
+
 ### 0.6.0
 
-- Region
+- rename to region-core and region-shortcut
 
-- sideEffect
+- Region
 
 - getProvider